authentication/requirements.txt
FLASK == 2.3.2
FLASK_CORS == 4.0.0
FLASK_SQLALCHEMY == 3.0.5
FLASK_LOGIN == 0.6.2
PYTHON-DOTENV == 1.0.0
PATH == 16.7.1
PSYCOPG2 == 2.9.7
FLASK_JWT_EXTENDED